+++ src/Http/Api/Contracts/HasPolicies.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\Gate;
 
 trait HasPolicies
 {
+
     /**
      * Qualifies the collection query to allow you to add params vai the policy
      * ie to limit to a specific user id mapping.